pyproject.toml
[project]
name = "procurement-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.25.0",
    "black>=24.0.0",
    "dspy-ai>=2.5.0",
    "faker>=38.2.0",
    "jupyter>=1.0.0",
    "mypy>=1.8.0",
    "notebook>=7.0.0",
    "openai>=1.0.0",
    "pandas>=2.3.3",
    "pymilvus[model]>=2.6.4",
    "pyyaml>=6.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.3.0",
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = ["__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E203", "E501"]

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
addopts = ["-q"]
